requirements.txt
opencv-python
pillow
mediapipe
pandas